feat: Move Connection-thunk logic to useConnect
